backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-mqtt==1.1.1
paho-mqtt==1.5.1
marshmallow==3.13.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
gunicorn==20.1.0
pytest==6.2.5
pytest-flask==1.2.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
apscheduler==3.8.1